Moved grammar.txt, imported formatter, used.
grammar.txt is now in the formatter/resources folder. learning-ring now
parses the code input into a tree, edits it, and spits it back out.

Unfortunately the method I used to load the files in formatter doesn't
work because it looks for the folder and it's not a resource-based
solution. So, it doesn't actually read *any* of the extensions! I will fix
this next.