Further additions to landing.
Mostly playing around with the landing page. Added a bar for the changes,
an output window to which the modified text will go. When posts code,
places original on top and modified code on the bottom.

Note that if it cannot parse the input it will throw an exception which
will splat to the user. Not Great; if I ever actually put this up, I
should remember to change that.