Modified when extension to update comments.
The comments are now updated with the character at which the change was
made in the when extension. It does this by diff-ing the two strings
created from the before and after trees using clj-diff.

Unfortunately, this means that if I follow the same procedure for every
extension, it'll be rendering the trees to strings and diffing each and
every step. That strikes me as awkward and inelegant.