Experimenting with compojure.
compojure is a routing library which is apparently very popular.

Changed structure to reflect the routing system. There's no "home" though
- if I had one I guess I'd put it as routes/home.clj, and have that be a
special case - or maybe even pull it out of the routes directory entirely?

With this, I think I have a fair idea of how to hack together a website.
I'm considering deploying this - or something similar - to...somewhere.