Changed extensions to be statically referenced.
I was trying to set this up to try deploying it to heroku, but
unfortunately there's all kinds of Messiness that's involved in
dynamically scanning the jar for the extensions. In the end I decided that
if I'm just going to pack these things into a jar anyways, what's the
point of writing it so you can just drop in a new extension?

See, the original idea to reading the folder was that it would make it
easy to extend, but then instead of using it like an application I ended
up packing it into a jar.

So, if you need to add a new extension, you'll have to go into
extensions.clj and manually add it to the list of all extensions. I
suppose I could add in a step to auto-generate the code...but the project
isn't nearly complex enough to warrant that, yet.

This will obviously break the tests.