Initial thread-last extension.
There are some issues with thread-last, however. The first and most
pressing is that I'm not quite sure what I want the expected behavior to
be! For example, given (* (- 4 (- 2 3)) (* 9 (+ 4 (- 2 4)))), what
suggestions should be returned?

The current method is controlled by the factors - a list of functions for
which it will *not* attempt to thread (for example, it would make no sense
to try and thread an if statement), and a level of nesting to which it
should thread.