Updated when.
It passes the tests - for the first time in a long time all the tests I
have pass. That's pretty nice.

The implementation itself features a rather, uh, strange usage of mapcat,
but that was the only way I could think of to split one node into many
nodes. I'm pretty sure that's more a failure of my imagination than
anything else, and that there's totally a way to do it more idiomatically
which I've missed horribly.

On the other hand it works!