Added a utility function to run thread-first/last
Added function in core. This is mostly to make it easier to see what's
actually going on for me. Also added :name field to extensions
thread-first and thread-last (and then to the rest of them). :name should
be a uniquely identifying descriptive name pertaining to the function of
the extension.