Moved markov gen code to clojurescript.
Also, mucked around with the landing html a bit. There's now a buttton you
can hit to generate a sonnet (starting with "Time").

Unfortunately I don't know exactly how I want the mapping from user-text
to generated-text to look like. Should the generated-text change if the
user goes back and edits something, or should it just keep generating? How
do you generate the new text such that it doesn't have to regenerate the
whole document whenever an early edit is made? Should it generate a bunch
of short mappings, or...?