Trying to get corpus -> json of counts.
Unfortunately it's a leetle trickier than I thought it'd be. Also, I'd
rather the json actually have the strings in the maps, rather than the
lists of characters (it all ends up the same in Clojure, really), so I
need to turn they values into maps of strings instead of maps of seqs.