Added in the ability to pick your corpus.
Currently, the list of corpora is hardcoded, as is the arity of the keys.
Also, it's not coded particularly well.

Also, if you keep typing while it's getting the new json it'll generate
with the old json.

Also, if it caches an output for a sentence it doesn't regenerate it if
you switch out the corpus.

These need to be fixed one way or another.