dfs builder now runs dfs, limited to 1 element
Builder now has the functionality to run the dfs, but it won't work on
maps whose states are larger than one element. I plan to address, this, of
course, after I wrap it into a two-argument function.

Also, I need to handle the path if there is no sentence ending with a
terminal (in which case it should generate any sentence).