Attempting to diagnose SO.
I think that the SO is due to deep recursion - that is, that it's due to
using all the stack frames, not due to allocating all the stack memory. My
first suspicion (that it wasn't throwing away dead-end trees due to my
somehow misunderstanding my lazy sequences) was wrong, it seems.

In order to fix it I'll have to convert it to a continuation so it doesn't
spam the stack with a million calls.