Further diagnosing the problems with my alg.
It turns out that there are quite a lot of issues! Especially thorny is
what behavior is expected if the length of the sentence to be generated
larger than the longest possible sentence even if it does not end on an
end node (for example, a language of a, b, c, ... z where each letter only
leads to the next one and an end node, asked to generate 50-length
string).