Sort of fixed dfs.
Revised dive algorithm (uses backtracking now) and also fixed the dfs.
Unfortunately, it doesn't have any provision for the case where it
traverses the *entire* tree because no valid sentence can be found and
then dies horribly, but takes a long time doing it. So that's an Issue
(try generating a 153-length sentence with n=2 on loremipsum)