Initial work on compressing datamap.
Actually, compressing refers not to the size of the datamap in text form,
but in the depth. The intent is to change the datamap such that walking
the following:
  "he" -> "is" -> "very" -> "angry" -> [right or]
should be much shorter, of the form
  "he is very angry" -> [right or]
Unfortunately this would also mean I'd need to modify a lot of other code
to get it to properly recognize the new data structure, and I haven't yet
done an actual analysis of the performance implications (it could be it
would hardly help where you'd notice and wouldn't at all when you'd need
it) so I'm branching it. I may just go ahead and go back to the frontend,
and leave this for later.